+++ src/commonMain/kotlin/org.angproj.aux/pkg/prime/LongType.kt
@@ -27,14 +27,14 @@ import kotlin.jvm.JvmInline
 @JvmInline
 public value class LongType(public val value: Long) : Enfoldable {
 
-    override fun foldSize(foldFormat: FoldFormat): Long = atomicSize.toLong()
+    override fun foldSize(foldFormat: FoldFormat): Int = atomicSize
 
-    public fun enfoldToBlock(outData: Storable, offset: Int): Long {
+    public fun enfoldToBlock(outData: Storable, offset: Int): Int {
         outData.storeLong(offset, value)
         return foldSize(FoldFormat.BLOCK)
     }
 
-    public fun enfoldToStream(outStream: BinaryWritable): Long {
+    public fun enfoldToStream(outStream: BinaryWritable): Int {
         outStream.writeLong(value)
         return foldSize(FoldFormat.STREAM)
     }
@@ -43,6 +43,7 @@ public value class LongType(public val value: Long) : Enfoldable {
         override val foldFormatSupport: List<FoldFormat> = listOf(FoldFormat.BLOCK, FoldFormat.STREAM)
         override val conventionType: Convention = Convention.LONG
         override val atomicSize: Int = Long.SIZE_BYTES
+
         public fun unfoldFromBlock(inData: Retrievable, offset: Int): LongType = LongType(inData.retrieveLong(offset))
 
         public fun unfoldFromStream(inStream: BinaryReadable): LongType = LongType(inStream.readLong())
