+++ src/commonMain/kotlin/org.angproj.aux/pkg/prime/UIntType.kt
@@ -27,14 +27,14 @@ import kotlin.jvm.JvmInline
 @JvmInline
 public value class UIntType(public val value: UInt) : Enfoldable {
 
-    override fun foldSize(foldFormat: FoldFormat): Long = atomicSize.toLong()
+    override fun foldSize(foldFormat: FoldFormat): Int = atomicSize
 
-    public fun enfoldToBlock(outData: Storable, offset: Int): Long {
+    public fun enfoldToBlock(outData: Storable, offset: Int): Int {
         outData.storeUInt(offset, value)
         return foldSize(FoldFormat.BLOCK)
     }
 
-    public fun enfoldToStream(outStream: BinaryWritable): Long {
+    public fun enfoldToStream(outStream: BinaryWritable): Int {
         outStream.writeUInt(value)
         return foldSize(FoldFormat.STREAM)
     }
@@ -43,6 +43,7 @@ public value class UIntType(public val value: UInt) : Enfoldable {
         override val foldFormatSupport: List<FoldFormat> = listOf(FoldFormat.BLOCK, FoldFormat.STREAM)
         override val conventionType: Convention = Convention.UINT
         override val atomicSize: Int = UInt.SIZE_BYTES
+
         public fun unfoldFromBlock(inData: Retrievable, offset: Int): UIntType = UIntType(inData.retrieveUInt(offset))
 
         public fun unfoldFromStream(inStream: BinaryReadable): UIntType = UIntType(inStream.readUInt())
